README.md now tells user about <form> field and also adds syntax highlighing to docs
heya :wave:

An aspect of this plugin that wasn't clear for at first was that I needed to wrap my input field in `<form>` tag.  I think it might be good to be extra clear about the need for it in the docs, otherwise the plugin doesn't work as can be seen [here](https://github.com/swiftype/swiftype-search-jquery/blob/master/jquery.swiftype.search.js#L146-L148)

Also I added syntax highlighting to the docs and tighten up the spaces a bit for easier reading on smaller screens.